+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Added testing for Python 3.14
 - Added logging of the instrument serial number to GeoCOM connection
+- Added instrument info logging to GSI Online DNA initialization
 
 ### Changed